refactor: reduce AI call context validator complexity
